Move attachment creators to mailer class
