+++ lib/private/Mail/Message.php
@@ -45,27 +45,6 @@ class Message implements IMessage {
 	}
 
 	/**
-	 * @param string|null $data
-	 * @param string|null $filename
-	 * @param string|null $contentType
-	 * @return IAttachment
-	 * @since 13.0.0
-	 */
-	public function createAttachment($data = null, $filename = null, $contentType = null) {
-		return new Attachment(\Swift_Attachment::newInstance($data, $filename, $contentType));
-	}
-
-	/**
-	 * @param string $path
-	 * @param string|null $contentType
-	 * @return IAttachment
-	 * @since 13.0.0
-	 */
-	public function createAttachmentFromPath($path, $contentType = null) {
-		return new Attachment(\Swift_Attachment::fromPath($path, $contentType));
-	}
-
-	/**
 	 * @param IAttachment $attachment
 	 * @return $this
 	 * @since 13.0.0
